Fix issue #2 by supporting Android on the example
