+++ index.android.js
@@ -1,52 +1,10 @@
 /**
- * Sample React Native App
- * https://github.com/facebook/react-native
+ * Sample Firebase & React Native App
+ * https://github.com/davideast/firebase-react-native-sample
  */
 'use strict';
 
-var React = require('react-native');
-var {
-  AppRegistry,
-  StyleSheet,
-  Text,
-  View,
-} = React;
-
-var FirebaseReactNativeSample = React.createClass({
-  render: function() {
-    return (
-      <View style={styles.container}>
-        <Text style={styles.welcome}>
-          Welcome to React Native!
-        </Text>
-        <Text style={styles.instructions}>
-          To get started, edit index.android.js
-        </Text>
-        <Text style={styles.instructions}>
-          Shake or press menu button for dev menu
-        </Text>
-      </View>
-    );
-  }
-});
-
-var styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#F5FCFF',
-  },
-  welcome: {
-    fontSize: 20,
-    textAlign: 'center',
-    margin: 10,
-  },
-  instructions: {
-    textAlign: 'center',
-    color: '#333333',
-    marginBottom: 5,
-  },
-});
+import { AppRegistry } from 'react-native';
+import FirebaseReactNativeSample from './App';
 
 AppRegistry.registerComponent('FirebaseReactNativeSample', () => FirebaseReactNativeSample);
